pyproject.toml
[project]
name = "story-book-creator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "asyncio>=3.4.3",
    "google-generativeai>=0.8.5",
    "loguru>=0.7.3",
    "openai>=1.76.2",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
]
